requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
SQLAlchemy==2.0.9
Werkzeug==2.2.3
Pillow==9.5.0
python-slugify==8.0.1
python-dotenv==1.0.0
email-validator==2.0.0
bleach==6.0.0
gunicorn==20.1.0
pyodbc==4.0.39
alembic==1.10.4
Jinja2==3.1.2
Markdown==3.4.3
WTForms==3.0.1
Flask-Compress==1.17